+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [2.0.0] - 2024-05-12
+## [2.0.0] - 2024-05-18
 
 General:
   - Add: New fonts, colours and general branding
@@ -27,6 +27,9 @@ Sidebar:
 
 Events:
   - Add: All new Events popups
+
+Community:
+  - Add: New community page that encourages interaction before getting Discord link
 
 Pages:
   - Add: New global styling for pages